#88 Add GraphQL support
