+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Features
 * [#81] - Improve cache & introduce such for query API
 * [#83] - Migrate JUnit 4 -> JUnit 5
+* [#88] - Add GraphQL support
 
 ## 1.2.1
 ### Miscellaneous
