+++ metadata-engine/src/main/java/io/github/sergeivisotsky/metadata/engine/config/EngineConfig.java
@@ -25,5 +25,5 @@ import org.springframework.web.bind.annotation.ControllerAdvice;
  */
 @Configuration
 @ConditionalOnMissingBean(annotation = ControllerAdvice.class)
-public class SelectorConfig {
+public class EngineConfig {
 }
