+++ metadata-engine/src/main/resources/META-INF/spring.factories
@@ -2,7 +2,7 @@ org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
   io.github.sergeivisotsky.metadata.engine.config.MetadataSelectorConfig,\
   io.github.sergeivisotsky.metadata.engine.config.DataSourceConfig,\
   io.github.sergeivisotsky.metadata.engine.config.DataSourceDialectConfig,\
-  io.github.sergeivisotsky.metadata.engine.config.SelectorConfig,\
+  io.github.sergeivisotsky.metadata.engine.config.EngineConfig,\
   io.github.sergeivisotsky.metadata.engine.config.ChartMetadataConfig,\
   io.github.sergeivisotsky.metadata.engine.config.PropertiesConfig,\
   io.github.sergeivisotsky.metadata.engine.config.CacheableMetadataConfig,\
