+++ metadata-graphql/pom.xml
@@ -1,0 +1,98 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>io.github.sergeivisotsky.metadata</groupId>
+        <artifactId>metadata-parent</artifactId>
+        <version>${revision}</version>
+    </parent>
+
+    <artifactId>metadata-graphql</artifactId>
+
+    <dependencies>
+        <dependency>
+            <groupId>io.github.sergeivisotsky.metadata</groupId>
+            <artifactId>metadata-engine</artifactId>
+            <version>${project.parent.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.junit.jupiter</groupId>
+            <artifactId>junit-jupiter-engine</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.mockito</groupId>
+            <artifactId>mockito-junit-jupiter</artifactId>
+            <scope>test</scope>
+        </dependency>
+    </dependencies>
+
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-surefire-plugin</artifactId>
+            </plugin>
+            <plugin>
+                <groupId>org.jacoco</groupId>
+                <artifactId>jacoco-maven-plugin</artifactId>
+            </plugin>
+            <plugin>
+                <groupId>io.github.kobylynskyi</groupId>
+                <artifactId>graphql-codegen-maven-plugin</artifactId>
+                <executions>
+                    <execution>
+                        <goals>
+                            <goal>generate</goal>
+                        </goals>
+                        <configuration>
+                            <graphqlSchemaPaths>
+                                <graphqlSchemaPath>
+                                    ${project.basedir}/src/main/resources/graphql/view_metadata.graphqls
+                                </graphqlSchemaPath>
+                                <graphqlSchemaPath>
+                                    ${project.basedir}/src/main/resources/graphql/navigation_metadata.graphqls
+                                </graphqlSchemaPath>
+                                <graphqlSchemaPath>
+                                    ${project.basedir}/src/main/resources/graphql/view_query.graphqls
+                                </graphqlSchemaPath>
+                                <graphqlSchemaPath>
+                                    ${project.basedir}/src/main/resources/graphql/lookup.graphqls
+                                </graphqlSchemaPath>
+                                <graphqlSchemaPath>
+                                    ${project.basedir}/src/main/resources/graphql/form_metadata.graphqls
+                                </graphqlSchemaPath>
+                                <graphqlSchemaPath>
+                                    ${project.basedir}/src/main/resources/graphql/view_metadata.graphqls
+                                </graphqlSchemaPath>
+                                <graphqlSchemaPath>
+                                    ${project.basedir}/src/main/resources/graphql/enum.graphqls
+                                </graphqlSchemaPath>
+                                <graphqlSchemaPath>
+                                    ${project.basedir}/src/main/resources/graphql/chart_metadata.graphqls
+                                </graphqlSchemaPath>
+                            </graphqlSchemaPaths>
+                            <outputDir>${project.build.directory}/generated-sources/graphql</outputDir>
+                            <apiPackageName>io.github.sergeivisotsky.metadata.graphql.api</apiPackageName>
+                            <modelPackageName>io.github.sergeivisotsky.metadata.graphql.model</modelPackageName>
+                            <customTypesMapping>
+                                <DateTime>java.time.LocalDateTime</DateTime>
+                            </customTypesMapping>
+                            <modelNameSuffix>TO</modelNameSuffix>
+                            <!--<parentInterfaces>
+                                <queryResolver>graphql.kickstart.tools.GraphQLQueryResolver</queryResolver>
+                                <mutationResolver>graphql.kickstart.tools.GraphQLMutationResolver</mutationResolver>
+                                <subscriptionResolver>graphql.kickstart.tools.GraphQLSubscriptionResolver</subscriptionResolver>
+                                <resolver><![CDATA[graphql.kickstart.tools.GraphQLResolver<{{TYPE}}>]]></resolver>
+                            </parentInterfaces>-->
+                        </configuration>
+                    </execution>
+                </executions>
+            </plugin>
+        </plugins>
+    </build>
+
+</project>
