+++ metadata-graphql/src/main/resources/graphql/scalars.graphqls
@@ -1,0 +1,3 @@
+scalar Object
+scalar Long
+scalar Double
