+++ metadata-rest/src/main/java/io/github/sergeivisotsky/metadata/engine/rest/ChartMetadataController.java
