+++ metadata-rest/src/main/java/io/github/sergeivisotsky/metadata/engine/rest/ViewMetadataController.java
