+++ metadata-rest/src/main/java/io/github/sergeivisotsky/metadata/engine/rest/dto/ChartMetadataRequest.java
