+++ metadata-rest/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+  io.github.sergeivisotsky.metadata.engine.rest.config.ChartMetadataRestConfig
